seer-backend/pyproject.toml
[project]
name = "seer-backend"
version = "0.1.0"
description = "Seer backend API scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<1",
  "httpx>=0.28.0,<1",
  "pydantic-settings>=2.11.0,<3",
  "pyshacl>=0.30.1,<1",
  "rdflib>=7.1.0,<8",
  "xxhash>=3.6.0,<4",
  "uvicorn[standard]>=0.35.0,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9",
  "ruff>=0.13.0,<1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
